Shorten Create Wiki (#196)
