+++ src/pages/create-wiki/index.tsx
@@ -1,10 +1,8 @@
 import React, {
   useContext,
   useEffect,
-  useState,
   useRef,
   memo,
-  useCallback,
   ChangeEvent,
   useMemo,
 } from 'react'
@@ -40,154 +38,87 @@ import {
   getRunningOperationPromises,
   getWiki,
   postWiki,
-  useGetWikiQuery,
 } from '@/services/wikis'
 import { useRouter } from 'next/router'
 import { store } from '@/store/store'
 import { GetServerSideProps } from 'next'
-import { skipToken } from '@reduxjs/toolkit/query'
-import { useAccount, useSignTypedData, useWaitForTransaction } from 'wagmi'
+import { useAccount } from 'wagmi'
 import { MdTitle } from 'react-icons/md'
 import slugify from 'slugify'
-import axios from 'axios'
-import diff from 'fast-diff'
 
 import Highlights from '@/components/Layout/Editor/Highlights/Highlights'
-import config from '@/config'
-import { useAppDispatch, useAppSelector } from '@/store/hook'
+import { useAppSelector } from '@/store/hook'
 import { getWikiMetadataById } from '@/utils/getWikiFields'
-import { PageTemplate } from '@/data/pageTemplate'
 import { getDeadline } from '@/utils/getDeadline'
-import { submitVerifiableSignature } from '@/utils/postSignature'
 import { ImageContext, ImageKey, ImageStateType } from '@/context/image.context'
 import { authenticatedRoute } from '@/components/AuthenticatedRoute'
 import WikiProcessModal from '@/components/Elements/Modal/WikiProcessModal'
 import { getWordCount } from '@/utils/getWordCount'
-import { POST_IMG } from '@/services/wikis/queries'
-import {
-  MData,
-  Wiki,
-  CommonMetaIds,
-  EditSpecificMetaIds,
-  WikiRootBlocks,
-} from '@/types/Wiki'
+import { Wiki, CommonMetaIds, EditSpecificMetaIds } from '@/types/Wiki'
 import { logEvent } from '@/utils/googleAnalytics'
+import {
+  initialMsg,
+  MINIMUM_WORDS,
+  useCreateWikiState,
+  saveImage,
+  calculateEditInfo,
+  CreateWikiProvider,
+  useGetSignedHash,
+  useCreateWikiEffects,
+  useCreateWikiContext,
+} from '@/utils/create-wiki'
 
 const Editor = dynamic(() => import('@/components/Layout/Editor/Editor'), {
   ssr: false,
 })
 
-const initialEditorValue = ` `
-const initialMsg =
-  'Your Wiki is being processed. It will be available on the blockchain soon.'
-const errorMessage = 'Oops, An Error Occurred. Wiki could not be created'
-const successMessage = 'Wiki has been created successfully.'
-
 const deadline = getDeadline()
 
-const domain = {
-  name: 'EP',
-  version: '1',
-  chainId: config.chainId,
-  verifyingContract: config.wikiContractAddress,
-}
-
-const types = {
-  SignedPost: [
-    { name: 'ipfs', type: 'string' },
-    { name: 'user', type: 'address' },
-    { name: 'deadline', type: 'uint256' },
-  ],
-}
-
-const MINIMUM_WORDS = 150
-
-const CreateWiki = () => {
+const CreateWikiContent = () => {
   const wiki = useAppSelector(state => state.wiki)
-  const dispatch = useAppDispatch()
-  const router = useRouter()
   const toast = useToast()
-  const { slug } = router.query
-  const { isLoading: isLoadingWiki, data: wikiData } = useGetWikiQuery(
-    typeof slug === 'string' ? slug : skipToken,
-    {
-      skip: router.isFallback,
-    },
-  )
+
   const { image, ipfsHash, updateImageState, isWikiBeingEdited } =
     useContext<ImageStateType>(ImageContext)
   const [{ data: accountData }] = useAccount()
-  const [md, setMd] = useState<string>()
-  const [openTxDetailsDialog, setOpenTxDetailsDialog] = useState<boolean>(false)
-  const [isWritingCommitMsg, setIsWritingCommitMsg] = useState<boolean>(false)
-  const [txHash, setTxHash] = useState<string>()
-  const [submittingWiki, setSubmittingWiki] = useState(false)
-  const [wikiHash, setWikiHash] = useState<string>()
-  const [isNewCreateWiki, setIsNewCreateWiki] = useState<boolean>(false)
-  const [activeStep, setActiveStep] = useState<number>(0)
-  const [loadingState, setIsLoading] = useState<
-    'error' | 'loading' | undefined
-  >('loading')
-  const [wikiId, setWikiId] = useState<string>('')
-  const [msg, setMsg] = useState<string>(initialMsg)
-  const [txError, setTxError] = useState({
-    title: '',
-    description: '',
-    opened: false,
-  })
-  const [{ data, error, loading: signing }, signTypedData] = useSignTypedData(
-    {},
-  )
-  const [, wait] = useWaitForTransaction()
+
+  const {
+    isLoadingWiki,
+    wikiData,
+    dispatch,
+    md,
+    setMd,
+    openTxDetailsDialog,
+    setOpenTxDetailsDialog,
+    isWritingCommitMsg,
+    setIsWritingCommitMsg,
+    txHash,
+    submittingWiki,
+    setSubmittingWiki,
+    wikiHash,
+    isNewCreateWiki,
+    activeStep,
+    setActiveStep,
+    loadingState,
+    setIsLoading,
+    wikiId,
+    setWikiId,
+    msg,
+    setMsg,
+    txError,
+    setTxError,
+  } = useCreateWikiContext()
+
   const prevEditedWiki = useRef<{ wiki?: Wiki; isPublished: boolean }>({
     wiki: wikiData,
     isPublished: false,
   })
 
-  const saveImage = async () => {
-    const formData = new FormData()
-    const blob = new Blob([image.type], {
-      type: 'multipart/form-data',
-    })
+  const { saveHashInTheBlockchain, signing, verifyTrxHash } =
+    useGetSignedHash(deadline)
 
-    formData.append('operations', POST_IMG)
-    const map = `{"0": ["variables.file"]}`
-    formData.append('map', map)
-    formData.append('0', blob)
-
-    const {
-      data: {
-        data: {
-          pinImage: { IpfsHash },
-        },
-      },
-    } = await axios.post(config.graphqlUrl, formData, {})
-
-    return IpfsHash
-  }
-
-  const saveHashInTheBlockchain = async (ipfs: string) => {
-    setWikiHash(ipfs)
-    signTypedData({
-      domain,
-      types,
-      value: {
-        ipfs,
-        user: accountData?.address || '',
-        deadline,
-      },
-    }).then(response => {
-      if (response.data) {
-        setActiveStep(1)
-      }
-      if (response.error) {
-        setIsLoading('error')
-        setMsg(errorMessage)
-      }
-    })
-  }
-
-  const getImageHash = async () => (isWikiBeingEdited ? ipfsHash : saveImage())
+  const getImageHash = async () =>
+    isWikiBeingEdited ? ipfsHash : saveImage(image)
 
   const getWikiSlug = () => slugify(String(wiki.title).toLowerCase())
 
@@ -241,93 +172,6 @@ const CreateWiki = () => {
 
     return true
   }
-  const calculateEditInfo = (prevWiki: Wiki, currWiki: Wiki) => {
-    const calculateContentChanged = () => {
-      // check if content has changed
-      const prevContent = prevWiki?.content
-      const currContent = currWiki?.content
-
-      // calculate percent changed and number of words changed in prevContent and currContent
-      let contentAdded = 0
-      let contentRemoved = 0
-      let contentUnchanged = 0
-
-      let wordsAdded = 0
-      let wordsRemoved = 0
-
-      diff(prevContent, currContent).forEach(part => {
-        if (part[0] === 1) {
-          contentAdded += part[1].length
-          wordsAdded += getWordCount(part[1])
-        }
-        if (part[0] === -1) {
-          contentRemoved += part[1].length
-          wordsRemoved += getWordCount(part[1])
-        }
-        if (part[0] === 0) {
-          contentUnchanged += part[1].length
-        }
-      })
-
-      const percentChanged =
-        ((contentAdded + contentRemoved) / contentUnchanged) * 100
-      const wordsChanged = wordsAdded + wordsRemoved
-
-      // update metadata in redux state
-      dispatch({
-        type: 'wiki/updateMetadata',
-        payload: {
-          id: EditSpecificMetaIds.WORDS_CHANGED,
-          value: wordsChanged.toString(),
-        },
-      })
-
-      dispatch({
-        type: 'wiki/updateMetadata',
-        payload: {
-          id: EditSpecificMetaIds.PERCENT_CHANGED,
-          value: percentChanged.toFixed(2),
-        },
-      })
-    }
-
-    // calculate which blocks have changed
-    const blocksChanged = []
-
-    // root level block changes
-    if (prevWiki.content !== currWiki.content) {
-      blocksChanged.push(WikiRootBlocks.CONTENT)
-      calculateContentChanged()
-    }
-    if (prevWiki.title !== currWiki.title)
-      blocksChanged.push(WikiRootBlocks.TITLE)
-    if (prevWiki.categories !== currWiki.categories)
-      blocksChanged.push('categories')
-    if (prevWiki.tags !== currWiki.tags) blocksChanged.push(WikiRootBlocks.TAGS)
-    if (prevWiki.summary !== currWiki.summary)
-      blocksChanged.push(WikiRootBlocks.SUMMARY)
-    const prevImgId = prevWiki.images && prevWiki.images[0].id
-    const currImgId = currWiki.images && currWiki.images[0].id
-    if (prevImgId !== currImgId) blocksChanged.push(WikiRootBlocks.WIKI_IMAGE)
-
-    // common metadata changes
-    Object.values(CommonMetaIds).forEach(id => {
-      if (
-        getWikiMetadataById(prevWiki, id)?.value !==
-        getWikiMetadataById(currWiki, id)?.value
-      )
-        blocksChanged.push(id)
-    })
-
-    // update blocks changed metadata in redux state
-    dispatch({
-      type: 'wiki/updateMetadata',
-      payload: {
-        id: EditSpecificMetaIds.BLOCKS_CHANGED,
-        value: blocksChanged.join(','),
-      },
-    })
-  }
 
   const saveOnIpfs = async () => {
     if (!isValidWiki()) return
@@ -363,9 +207,9 @@ const CreateWiki = () => {
         // more than two edits to chain in same session
 
         if (prevEditedWiki.current.isPublished && prevEditedWiki.current.wiki) {
-          calculateEditInfo(prevEditedWiki.current.wiki, interWiki)
+          calculateEditInfo(prevEditedWiki.current.wiki, interWiki, dispatch)
         } else if (wikiData) {
-          calculateEditInfo(wikiData, interWiki)
+          calculateEditInfo(wikiData, interWiki, dispatch)
         }
       }
 
@@ -410,114 +254,13 @@ const CreateWiki = () => {
     setMd(val || ' ')
   }
 
-  const updatePageTypeTemplate = useCallback(() => {
-    const meta = [
-      getWikiMetadataById(wiki, CommonMetaIds.PAGE_TYPE),
-      getWikiMetadataById(wiki, CommonMetaIds.TWITTER_PROFILE),
-    ]
-    const pageType = PageTemplate.find(p => p.type === meta[0]?.value)
-
-    setMd(String(pageType?.templateText))
-  }, [wiki])
-
-  const verifyTrxHash = useCallback(
-    async (trxHash: string) => {
-      const timer = setInterval(() => {
-        try {
-          const checkTrx = async () => {
-            const trx = await wait({
-              hash: trxHash,
-            })
-            if (trx.error) {
-              setIsLoading('error')
-              setMsg(errorMessage)
-              clearInterval(timer)
-            } else if (trx.data.confirmations > 1) {
-              setIsLoading(undefined)
-              setActiveStep(3)
-              setMsg(successMessage)
-              clearInterval(timer)
-            }
-          }
-          checkTrx()
-        } catch (err) {
-          setIsLoading('error')
-          setMsg(errorMessage)
-          clearInterval(timer)
-        }
-      }, 3000)
-    },
-    [wait],
-  )
+  useCreateWikiEffects(wiki, prevEditedWiki)
 
   useEffect(() => {
     if (activeStep === 3) {
       prevEditedWiki.current.isPublished = true
     }
   }, [activeStep])
-
-  // Reset the State to new wiki if there is no slug
-  useEffect(() => {
-    if (!slug) {
-      setIsNewCreateWiki(true)
-      dispatch({ type: 'wiki/reset' })
-      setMd(initialEditorValue)
-    } else {
-      setIsNewCreateWiki(false)
-    }
-  }, [dispatch, slug])
-
-  useEffect(() => {
-    if (isLoadingWiki === false && !wikiData) setMd(initialEditorValue)
-  }, [isLoadingWiki, wikiData])
-
-  // update the page type template when the page type changes
-  const presentPageType = useMemo(
-    () =>
-      wiki?.metadata?.find((m: MData) => m.id === CommonMetaIds.PAGE_TYPE)
-        ?.value,
-    [wiki?.metadata],
-  )
-  useEffect(() => {
-    if (presentPageType) {
-      let isMdPageTemplate = false
-      PageTemplate.forEach(p => {
-        if (p.templateText === md) isMdPageTemplate = true
-      })
-      if (isMdPageTemplate || md === ' ') updatePageTypeTemplate()
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [presentPageType])
-
-  useEffect(() => {
-    const getSignedTxHash = async () => {
-      if (data && wikiHash && accountData) {
-        if (error) {
-          setMsg(errorMessage)
-          setIsLoading('error')
-          return
-        }
-        try {
-          const { data: relayerData } = await submitVerifiableSignature(
-            data,
-            wikiHash,
-            accountData?.address,
-            deadline,
-          )
-
-          if (relayerData && relayerData.hash) {
-            setTxHash(relayerData.hash)
-            setActiveStep(2)
-          }
-        } catch (err) {
-          setIsLoading('error')
-          setMsg(errorMessage)
-        }
-      }
-    }
-    getSignedTxHash()
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [data, error])
 
   useEffect(() => {
     if (
@@ -756,6 +499,17 @@ const CreateWiki = () => {
         </Flex>
       </Skeleton>
     </Box>
+  )
+}
+
+const CreateWiki = () => {
+  const router = useRouter()
+  const wikiState = useCreateWikiState(router)
+  const providerValue = useMemo(() => wikiState, [wikiState])
+  return (
+    <CreateWikiProvider value={providerValue}>
+      <CreateWikiContent />
+    </CreateWikiProvider>
   )
 }
